Fix incorrect spelling counting toward loop
I successfully used recursion to have the user prompted to re-enter
their input, being notified of their spelling error, without it counting
toward the loop count.

The game is still a little messy, but it's a nice start for Javascript.

For example, typing out your selection each time is cumbersome. It
could be redesigned to check a box or enter numbers 1-3 instead. Or
something similar.